+++ BreezeGameEngine/Beveler.h
@@ -8,10 +8,10 @@ class Beveler
 public:
 	Beveler() = default;
 	Beveler(Color base);
-	void DrawOuterBev(Rect<int>& Rect, int size, Graphics& gfx);
-	void DrawInnerBev(Rect<int>& Rect, int size, Graphics& gfx);
-	void DrawBevBrick(Rect<int>& Rect, int size, Graphics& gfx);
-	void DrawBevBorder(Rect<int>& Rect, int size, Graphics& gfx);
+	void DrawOuterBev(Rect<float>& Rect, int size, Graphics& gfx);
+	void DrawInnerBev(Rect<float>& Rect, int size, Graphics& gfx);
+	void DrawBevBrick(Rect<float>& Rect, int size, Graphics& gfx);
+	void DrawBevBorder(Rect<float>& Rect, int size, Graphics& gfx);
 	void ChangeBaseColor(Color newbase);
 private:
 	static constexpr float LeftFactor = 1.10f;
